Rename the Chewy renderer
