+++ lib/msfl_visitors.rb
@@ -1,7 +1,7 @@
 require_relative 'msfl_visitors/ast'
 require_relative 'msfl_visitors/nodes'
 require_relative 'msfl_visitors/visitor'
-require_relative 'msfl_visitors/renderers/chewy_renderer'
+require_relative 'msfl_visitors/renderers/chewy/term_filter'
 require_relative 'msfl_visitors/parsers/msfl_parser'
 
 module MSFLVisitors
